PyXFEpubReader/requirements.txt
PyQt6
beautifulsoup4
websocket-client
pyaudio